int-double fix

+++ Bezinkut.cs
@@ -1,18 +1,18 @@
 namespace ceg_auto;
 
-class Benzinkut(int benzinMennyiseg, int literAr, int kasszaPenz) {
-   public int BenzinMennyiseg { get; set; } = benzinMennyiseg;
+class Benzinkut(double benzinMennyiseg, int literAr, int kasszaPenz) {
+   public double BenzinMennyiseg { get; set; } = benzinMennyiseg;
    public int LiterAr { get; set; } = literAr;
    public int KasszaPenz { get; set; } = kasszaPenz;
 
-   public void Tankol(Auto auto, int liter) {
+   public void Tankol(Auto auto, double liter) {
       if(liter > BenzinMennyiseg) return;
       if (auto.BenzinSzint + liter > auto.TankMeret)
          liter = auto.TankMeret - auto.BenzinSzint;
 
       auto.BenzinSzint += liter;
       BenzinMennyiseg -= liter;
-      auto.SoforPenz -= liter * LiterAr;
-      KasszaPenz += liter * LiterAr;
+      auto.SoforPenz -= (int)(liter * LiterAr);
+      KasszaPenz += (int)(liter * LiterAr);
    }
 }
